`model`, `sampler`, `spacing` in `build_jobs()`
They're now part of the job system like any other parameter.

`model` is manually zipped in at the end to ensure the least pipe swaps.
Currenlty this means it doesn't respect `--iter`

There is unfortunately manual GC calls because `del` won't trigger until
what I assume is the end of scope.

`build_schedulers` reworked to `get_scheduler`;
Scheduelrs are now built JIT for each job.
This is necessary because the default scheduler cannot be known until
the job is processed.