Refactor `--attention` -> `--attn-patch`
Attention processors are now based on patching the torch sdpa function
because Diffusers is extremely inconsistent with having a
`set_attn_processor` function for its structures.

Currently implemented a smarter version of the Flash for use with Navi3
users or whomever else decides they want to use a Flash Attention wheel.

Triton wheel support is written but untested as the Triton op segfaults
on my card/build so it's replaced with the DaoLab script

Remobed subquad entirely as it's not useful without extensive tuning for
an individual GPU.